+++ api/go.mod
@@ -18,9 +18,11 @@ require (
 	github.com/redis/go-redis/v9 v9.0.3 // indirect
 	golang.org/x/crypto v0.8.0 // indirect
 	golang.org/x/exp v0.0.0-20230419192730-864b3d6c5c2c // indirect
+	golang.org/x/mod v0.8.0 // indirect
 	golang.org/x/sync v0.1.0 // indirect
 	golang.org/x/sys v0.7.0 // indirect
 	golang.org/x/text v0.9.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
 replace bringyour.com/bringyour v0.0.0 => ../bringyour
