+++ bringyour/go.mod
@@ -11,6 +11,8 @@ require (
 	github.com/redis/go-redis/v9 v9.0.3
 	golang.org/x/crypto v0.8.0
 	golang.org/x/exp v0.0.0-20230419192730-864b3d6c5c2c
+	golang.org/x/mod v0.8.0
+	gopkg.in/yaml.v3 v3.0.1
 )
 
 require (
@@ -21,6 +23,8 @@ require (
 	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
 	github.com/jackc/puddle/v2 v2.2.0 // indirect
 	github.com/jmespath/go-jmespath v0.4.0 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/rogpeppe/go-internal v1.10.0 // indirect
 	golang.org/x/sync v0.1.0 // indirect
 	golang.org/x/sys v0.7.0 // indirect
 	golang.org/x/text v0.9.0 // indirect
